Refactor code (make multiple "helper" files)
